4.21

+++ part4/blogilista/controllers/blogs.js
@@ -7,10 +7,11 @@ const jwt = require('jsonwebtoken')
 const getTokenFrom = request => {
     const authorization = request.get('authorization')
     if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
+      console.log('getting token')
+      return authorization.substring(7)
     }
     return null
-}
+  }
 
 blogsRouter.get('/', async (request, response, next) => {
     try {
@@ -25,7 +26,6 @@ blogsRouter.get('/', async (request, response, next) => {
 blogsRouter.post('/', async (request, response, next) => {
     try {
         const token = getTokenFrom(request)
-
         const decodedToken = jwt.verify(token, process.env.SECRET)
         if (!token || !decodedToken.id) {
             return response.status(401).json({ error: 'token missing or invalid' })
@@ -54,7 +54,19 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try {
-        const blog = await Blog.findByIdAndRemove(request.params.id)
+        const token = getTokenFrom(request)
+
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(request.params.id)
+        if (user._id.toString() === blog.user.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+        } else {
+            return response.status(401).json({ error: 'not authorized' })
+        }
         response.status(204).end()
     } catch (exception) {
         next(exception)
